+++ lib/src/auth/auth.dart
@@ -1,0 +1,7 @@
+library flutter_dashboard_auth;
+
+import 'package:flutter/material.dart';
+import 'package:flutter_dashboard/flutter_dashboard.dart';
+
+part './auth_controller.dart';
+part './login_view.dart';